+++ src/commands/fun/bipen.ts
@@ -1,7 +1,6 @@
 import { Message, MessageEmbed } from "discord.js";
 import { BIPEN_IMG_URL } from "../../constants";
-import { Command } from "../../interfaces/command";
-import { CommandConfig } from "../../types/types";
+import { Command, CommandConfig } from "../../types/command";
 import { sendEmbeds } from "../../util";
 
 const cmdConfig: CommandConfig = {
