+++ src/commands/fun/scramble.ts
@@ -1,6 +1,5 @@
 import { Guild, Message, StageChannel, VoiceChannel } from "discord.js";
-import { Command } from "../../interfaces/command";
-import { CommandConfig } from "../../types/types";
+import { Command, CommandConfig } from "../../types/command";
 import { getRandomElement, sendErrorMessage, sendMessage } from "../../util";
 
 const cmdConfig: CommandConfig = {
