+++ src/commands/utility/connect.ts
@@ -1,7 +1,6 @@
 import { joinVoiceChannel } from "@discordjs/voice";
 import { Message } from "discord.js";
-import { Command } from "../../interfaces/command";
-import { CommandConfig } from "../../types/types";
+import { Command, CommandConfig } from "../../types/command";
 import { sendErrorMessage, sendMessage } from "../../util";
 
 const cmdConfig: CommandConfig = {
