+++ src/commands/utility/list_roles.ts
@@ -1,6 +1,5 @@
 import { Message, MessageEmbed } from "discord.js";
-import { Command } from "../../interfaces/command";
-import { CommandConfig } from "../../types/types";
+import { Command, CommandConfig } from "../../types/command";
 import { sendEmbeds } from "../../util";
 
 const cmdConfig: CommandConfig = {
