+++ src/commands/utility/ping.ts
@@ -1,6 +1,5 @@
 import { Message } from "discord.js";
-import { Command } from "../../interfaces/command";
-import { CommandConfig } from "../../types/types";
+import { Command, CommandConfig } from "../../types/command";
 
 const cmdConfig: CommandConfig = {
 	name: "ping",
