--- src/types/command_config.ts
@@ -1,11 +1,0 @@
-export interface CommandConfig {
-	name: string;
-	description: string;
-	usage: string;
-	cooldownTime?: number;
-	examples?: string[];
-	allowInDM?: boolean;
-	aliases?: string[];
-	disabled?: boolean;
-	permissions?: any[];
-}
