+++ src/util/index.ts
@@ -1,7 +1,6 @@
 export * from "./random";
-export * from "./schedule_utils";
 export * from "./sleep";
-export * from "./printSpace";
-export * from "./isDevMode";
-export * from "./messageChannel";
-export * from "./deleteVoiceChannel";
+export * from "./print_space";
+export * from "./is_dev_mode";
+export * from "./message_channel";
+export * from "./delete_voice_channel";
